requirements.txt
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
pytest
pandas
matplotlib
seaborn
openpyxl
click
tabulate
python-multipart
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt


